refactor(e2e): inline get-port and move catalog utils per app
Drop e2e-utils json/po catalog and playwright-dev-port; each Playwright
suite uses get-port directly (env shared when config is re-evaluated) and
workers: 1. Catalog helpers live beside their specs.
